requirements.txt
atproto
python-dotenv
shortuuid
boto3
botocore